Add documentation for @KeepGeneratedSerializer feature
Resolves #2660
